% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overloaded_extract_data_frame.R
\name{-}
\alias{-}
\title{Minus Operator}
\usage{
"-"(x, y)
}
\arguments{
\item{x, y}{numeric or complex vectors or objects which can be coerced to such, or other objects for which methods have been written.}
}
\description{
minus
}
