% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overloaded_extract_data_frame.R
\name{[}
\alias{[}
\title{Extract or Replace Parts of an Object}
\usage{
"["(x, i, j, drop = if (missing(i)) TRUE else length(cols) == 1)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i, j}{indices specifying elements to extract or replace. Indices are numeric or character vectors}

\item{drop}{See \code{\link[base]{drop}}.}
}
\description{
Operators acting on vectors, matrices, arrays and lists to extract or replace parts.
}
